% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoodmanKruskal.R
\docType{package}
\name{GoodmanKruskal}
\alias{GoodmanKruskal}
\alias{GoodmanKruskal-package}
\title{GoodmanKruskal: The Goodman and Kruskal tau measure}
\description{
Association analysis between categorical variables using the Goodman and Kruskal tau measure
}

