% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupNumericDataframe.R
\name{GroupNumericDataframe}
\alias{GroupNumericDataframe}
\title{Group numeric variables in a dataframe}
\usage{
GroupNumericDataframe(df, groupList)
}
\arguments{
\item{df}{Original dataframe from which new dataframe will
be created.}

\item{groupList}{List of parameters that specify which
numeric variables are to be grouped and the parameters
used for grouping them; see Details.}
}
\value{
Dataframe with numeric variables specified by
groupList grouped and converted to factors.
}
\description{
TODO: include description
}
\author{
Ron Pearson
}

