% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDataframe.R
\name{MakeDataframe}
\alias{MakeDataframe}
\title{Prepare a dataframe for the GKtauDataframe function}
\usage{
MakeDataframe(df, keepVars, recordSubset = NULL, groupNumericList = NULL,
  regroupCategoricalList = NULL)
}
\arguments{
\item{df}{Original dataframe from which new dataframe will
be created.}

\item{keepVars}{Character vector with names of variables from
df to be used in constructing the new dataframe.}

\item{recordSubset}{Integer vector of record indices to be
retained in constructing the new dataframe; if NULL (the
default), retain all records from df.}

\item{groupNumericList}{List of parameters to be passed to
the GroupNumericDataframe function.}

\item{regroupCategoricalList}{List of parameters to be
passed to the RegroupCategoricalDataframe function.}
}
\value{
Modified dataframe, constructed from the variables
specified in keepVars from df, the records specified by
recordSubset, and any data conversions specified by
groupNumericList and/or regroupCategoricalList.
}
\description{
This function prepares a dataframe for the GKtauDataframe
function, in four ways: (1), selecting a subset of variables
to retain for analysis; (2), extracting a subset of records;
(3), grouping numeric variables via the GroupNumeric function;
and (4), regrouping categorical variables via the function
RegroupCategorical.
}
\details{
This is a helper function to simplify the task of creating
dataframes for analysis using the GKtauDataframe function,
particularly when these results are to be plotted using the
plot.GKtauMatrix method.  The sequence of operations is this:
first, variables are selected according to keepVars and
records are retained as determined by recordSubset (if
this value is NULL, the default, all records are retained);
then, the function GroupNumericDataframe is applied using
the parameters specified by groupNumericList (if not NULL)
to create a dataframe with the specified numeric variables
grouped as factors; finally, the RegroupCategoricalDataframe
function is called to regroup categorical variables in this
modified dataframe using the parameters specfied by
regroupCategoricalList (if not NULL).
}
\author{
Ron Pearson
}

