% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegroupCategoricalDataframe.R
\name{RegroupCategoricalDataframe}
\alias{RegroupCategoricalDataframe}
\title{Regroup categorical variables in a dataframe}
\usage{
RegroupCategoricalDataframe(df, regroupList)
}
\arguments{
\item{df}{Original dataframe from which new dataframe will
be created.}

\item{regroupList}{List of parameters that specify which
categorical variables are to be regrouped and the parameters
used for regrouping them; see Details.}
}
\value{
Dataframe with the categorical variables specified by
regroupList regrouped.
}
\description{
TODO: include description
}
\author{
Ron Pearson
}

